+++ src/main/java/br/com/sistemapedidos/domain/entrega/FormaEntrega.java
@@ -2,5 +2,5 @@ package br.com.sistemapedidos.domain.entrega;
 
 public interface FormaEntrega {
     String entrega();
-    int calcularFrete();
+    double calcularFrete();
 }
